test/ingress: refactor and fix bug
This refactors the HTTP access code so it can be reused later for the
Headlamp ingress test.
