+++ test/ingress/aws/aws_test.go
@@ -18,22 +18,9 @@
 package aws
 
 import (
-	"context"
-	"crypto/tls"
-	"crypto/x509"
-	"fmt"
-	"net/http"
 	"testing"
-	"time"
 
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/util/wait"
-
-	testutil "github.com/kinvolk/lokomotive/test/components/util"
-)
-
-const (
-	httpTimeout = 4 * time.Second
+	testingress "github.com/kinvolk/lokomotive/test/ingress"
 )
 
 func TestAWSIngress(t *testing.T) {
@@ -61,89 +48,7 @@ func TestAWSIngress(t *testing.T) {
 		tc := tc
 		t.Run(tc.Ingress, func(t *testing.T) {
 			t.Parallel()
-
-			client := testutil.CreateKubeClient(t)
-
-			i, err := client.NetworkingV1beta1().Ingresses(tc.Namespace).Get(context.TODO(), tc.Component, metav1.GetOptions{})
-			if err != nil {
-				t.Fatalf("getting %s ingress: %v", tc.Component, err)
-			}
-
-			h := i.Spec.Rules[0].Host
-			c := getHTTPClient()
-
-			err = wait.PollImmediate(testutil.RetryInterval, testutil.TimeoutSlow, func() (bool, error) {
-				resp, err := c.Get(fmt.Sprintf("https://%s/%s", h, tc.Subpath))
-				if err != nil {
-					t.Logf("got an HTTP error: %v", err)
-					return false, nil
-				}
-
-				defer func() {
-					if err := resp.Body.Close(); err != nil {
-						t.Logf("closing HTTP response body: %v", err)
-					}
-				}()
-
-				if resp.StatusCode != http.StatusOK {
-					t.Logf("got a non-OK HTTP status: %d", resp.StatusCode)
-					return false, nil
-				}
-
-				return true, nil
-			})
-			if err != nil {
-				t.Fatal("could not get a successful HTTP response in time")
-			}
+			testingress.AccessIngress(t, tc.Namespace, tc.Ingress, tc.Subpath)
 		})
-	}
-}
-
-// getHTTPClient creates a HTTP client with LetsEncrypt Staging Root PEM certificate.
-func getHTTPClient() *http.Client {
-	// Get this Root PEM from https://letsencrypt.org/docs/staging-environment/#root-certificate
-	letsEncryptStagingRootPEM := `-----BEGIN CERTIFICATE-----
-MIIFATCCAumgAwIBAgIRAKc9ZKBASymy5TLOEp57N98wDQYJKoZIhvcNAQELBQAw
-GjEYMBYGA1UEAwwPRmFrZSBMRSBSb290IFgxMB4XDTE2MDMyMzIyNTM0NloXDTM2
-MDMyMzIyNTM0NlowGjEYMBYGA1UEAwwPRmFrZSBMRSBSb290IFgxMIICIjANBgkq
-hkiG9w0BAQEFAAOCAg8AMIICCgKCAgEA+pYHvQw5iU3v2b3iNuYNKYgsWD6KU7aJ
-diddtZQxSWYzUI3U0I1UsRPTxnhTifs/M9NW4ZlV13ZfB7APwC8oqKOIiwo7IwlP
-xg0VKgyz+kT8RJfYr66PPIYP0fpTeu42LpMJ+CKo9sbpgVNDZN2z/qiXrRNX/VtG
-TkPV7a44fZ5bHHVruAxvDnylpQxJobtCBWlJSsbIRGFHMc2z88eUz9NmIOWUKGGj
-EmP76x8OfRHpIpuxRSCjn0+i9+hR2siIOpcMOGd+40uVJxbRRP5ZXnUFa2fF5FWd
-O0u0RPI8HON0ovhrwPJY+4eWKkQzyC611oLPYGQ4EbifRsTsCxUZqyUuStGyp8oa
-aoSKfF6X0+KzGgwwnrjRTUpIl19A92KR0Noo6h622OX+4sZiO/JQdkuX5w/HupK0
-A0M0WSMCvU6GOhjGotmh2VTEJwHHY4+TUk0iQYRtv1crONklyZoAQPD76hCrC8Cr
-IbgsZLfTMC8TWUoMbyUDgvgYkHKMoPm0VGVVuwpRKJxv7+2wXO+pivrrUl2Q9fPe
-Kk055nJLMV9yPUdig8othUKrRfSxli946AEV1eEOhxddfEwBE3Lt2xn0hhiIedbb
-Ftf/5kEWFZkXyUmMJK8Ra76Kus2ABueUVEcZ48hrRr1Hf1N9n59VbTUaXgeiZA50
-qXf2bymE6F8CAwEAAaNCMEAwDgYDVR0PAQH/BAQDAgEGMA8GA1UdEwEB/wQFMAMB
-Af8wHQYDVR0OBBYEFMEmdKSKRKDm+iAo2FwjmkWIGHngMA0GCSqGSIb3DQEBCwUA
-A4ICAQBCPw74M9X/Xx04K1VAES3ypgQYH5bf9FXVDrwhRFSVckria/7dMzoF5wln
-uq9NGsjkkkDg17AohcQdr8alH4LvPdxpKr3BjpvEcmbqF8xH+MbbeUEnmbSfLI8H
-sefuhXF9AF/9iYvpVNC8FmJ0OhiVv13VgMQw0CRKkbtjZBf8xaEhq/YqxWVsgOjm
-dm5CAQ2X0aX7502x8wYRgMnZhA5goC1zVWBVAi8yhhmlhhoDUfg17cXkmaJC5pDd
-oenZ9NVhW8eDb03MFCrWNvIh89DDeCGWuWfDltDq0n3owyL0IeSn7RfpSclpxVmV
-/53jkYjwIgxIG7Gsv0LKMbsf6QdBcTjhvfZyMIpBRkTe3zuHd2feKzY9lEkbRvRQ
-zbh4Ps5YBnG6CKJPTbe2hfi3nhnw/MyEmF3zb0hzvLWNrR9XW3ibb2oL3424XOwc
-VjrTSCLzO9Rv6s5wi03qoWvKAQQAElqTYRHhynJ3w6wuvKYF5zcZF3MDnrVGLbh1
-Q9ePRFBCiXOQ6wPLoUhrrbZ8LpFUFYDXHMtYM7P9sc9IAWoONXREJaO08zgFtMp4
-8iyIYUyQAbsvx8oD2M8kRvrIRSrRJSl6L957b4AFiLIQ/GgV2curs0jje7Edx34c
-idWw1VrejtwclobqNMVtG3EiPUIpJGpbMcJgbiLSmKkrvQtGng==
------END CERTIFICATE-----`
-
-	rootCAs := x509.NewCertPool()
-	if ok := rootCAs.AppendCertsFromPEM([]byte(letsEncryptStagingRootPEM)); !ok {
-		// This should fail in the developer testing itself.
-		panic("Failed to parse root certificate")
-	}
-
-	return &http.Client{
-		Timeout: httpTimeout,
-		Transport: &http.Transport{
-			TLSClientConfig: &tls.Config{
-				RootCAs: rootCAs,
-			},
-		},
 	}
 }
